src/file_manager/requirements.txt
appnope==0.1.4
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.2
azure-functions==1.20.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.21.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
git-filter-repo==2.45.0
google-api-core==2.19.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
idna==3.7
importlib_metadata==8.0.0
ipykernel==6.29.4
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msal==1.30.0
msal-extensions==1.2.0
nest-asyncio==1.6.0
numpy==2.0.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pinecone-client==4.1.1
pinecone-notebooks==0.0.2rc3
pinecone-plugin-assistant==0.1.3
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.3
requests==2.32.3
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.19.2
